reduce native header duplication
